refactor(core): extract IPC and single-instance service to core.services.ipc_service
